+++ src/app/public/modules/auth-http-client/local-endpoint-checker.ts
